feat: move tree_sitter trait to specific file
